LayoutTests/js/resize-array-assign-expected.txt

Tests that assignments into arrays behave properly after the array is resized.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 1 is 1
PASS successfullyParsed is true

TEST COMPLETE
